matrix construction for FVM method

+++ tests/test_matrix.cpp
@@ -17,7 +17,7 @@ TEST(matrix, construct_from_parent_only)
         matrix_type m{p};
         EXPECT_EQ(m.num_cells(), 1);
         EXPECT_EQ(m.size(), 3);
-        EXPECT_EQ(p.size(), 3);
+        EXPECT_EQ(p.size(), 3u);
 
         auto mp = m.p();
         EXPECT_EQ(mp[0], 0);
@@ -32,7 +32,7 @@ TEST(matrix, construct_from_parent_only)
         matrix_type m{std::move(p)};
         EXPECT_EQ(m.num_cells(), 1);
         EXPECT_EQ(m.size(), 3);
-        EXPECT_EQ(p.size(), 3);
+        EXPECT_EQ(p.size(), 3u);
         EXPECT_EQ(m.size(), 3);
 
         auto mp = m.p();
@@ -49,7 +49,7 @@ TEST(matrix, construct_from_parent_only)
         matrix_type m{p};
         EXPECT_EQ(m.num_cells(), 1);
         EXPECT_EQ(m.size(), 3);
-        EXPECT_EQ(p.size(), 3);
+        EXPECT_EQ(p.size(), 3u);
 
         auto mp = m.p();
         EXPECT_EQ(mp[0], 0);
@@ -64,7 +64,7 @@ TEST(matrix, construct_from_parent_only)
         matrix_type m{std::move(p)};
         EXPECT_EQ(m.num_cells(), 1);
         EXPECT_EQ(m.size(), 3);
-        EXPECT_EQ(p.size(), 0); // 0 implies moved from
+        EXPECT_EQ(p.size(), 0u); // 0 implies moved from
 
         auto mp = m.p();
         EXPECT_EQ(mp[0], 0);
@@ -98,7 +98,7 @@ TEST(matrix, solve)
             std::iota(p.begin()+1, p.end(), 0);
             matrix_type m{p};
 
-            EXPECT_EQ(m.size(), n);
+            EXPECT_EQ(m.size(), (int)n);
             EXPECT_EQ(m.num_cells(), 1);
 
             m.d()(memory::all) =  2;
